Model SetFB
Set FB code in model class gestopt. Hij werkt en maakt contact. Krijg nog wel result = false terug, dus dit moet ik nog bekijken
